Updated various public files
Got rid of files that we weren't using anymore. Most of which were from the legacy days of the project when I was trying to build it all from scratch (luckly I wisened up and used an external editor). Also have made changes to the html file fixing small issues.